# Icon: optimize SVG with svgo