--- a/packages/gestalt/src/icons/cog.svg
+++ b/packages/gestalt/src/icons/cog.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M6.25 2L.5 12l5.75 10h11.5l5.75-10-5.75-10H6.25zm1.703 10a4 4 0 1 1 8 0 4 4 0 0 1-8 0z"/></svg>
+<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M6.25 2 .5 12l5.75 10h11.5l5.75-10-5.75-10H6.25zm1.703 10a4 4 0 1 1 8 0 4 4 0 0 1-8 0z"/></svg>