--- a/packages/gestalt/src/icons/dash.svg
+++ b/packages/gestalt/src/icons/dash.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M24 12a2 2 0 0 1-2 2H2a2 2 0 1 1 0-4h20a2 2 0 0 1 2 2z"/></svg>
+<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24"><path d="M24 12a2 2 0 0 1-2 2H2a2 2 0 1 1 0-4h20a2 2 0 0 1 2 2z"/></svg>