--- a/packages/gestalt/src/icons/dash.svg
+++ b/packages/gestalt/src/icons/dash.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M24 12a2 2 0 01-2 2H2a2 2 0 110-4h20a2 2 0 012 2z"/></svg>
+<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M24 12a2 2 0 0 1-2 2H2a2 2 0 1 1 0-4h20a2 2 0 0 1 2 2z"/></svg>