--- a/packages/gestalt/src/icons/flag.svg
+++ b/packages/gestalt/src/icons/flag.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M17 7.5L22 0H2v22c0 1.1.9 2 2 2s2-.9 2-2v-7h16l-5-7.5z"/></svg>
+<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M17 7.5 22 0H2v22c0 1.1.9 2 2 2s2-.9 2-2v-7h16l-5-7.5z"/></svg>