--- a/packages/gestalt/src/icons/flash.svg
+++ b/packages/gestalt/src/icons/flash.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M17.45 0L3.5 14h6.586L6.55 24 20.5 10h-6.586z"/></svg>
+<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M17.45 0 3.5 14h6.586L6.55 24 20.5 10h-6.586z"/></svg>