--- a/packages/gestalt/src/icons/gmail.svg
+++ b/packages/gestalt/src/icons/gmail.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M3.173 4 12 10.194 20.556 4H3.173zM12 13.29 1 5.806V20h22V5.548L12 13.29z"/></svg>
+<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24"><path d="M3.173 4 12 10.194 20.556 4H3.173zM12 13.29 1 5.806V20h22V5.548L12 13.29z"/></svg>