--- a/packages/gestalt/src/icons/svgo.config.js
+++ b/packages/gestalt/src/icons/svgo.config.js
@@ -1,0 +1,27 @@
+// @flow strict
+module.exports = {
+  plugins: [
+    {
+      name: 'preset-default',
+      //   params: {
+      //     overrides: {
+      //       // customize options for plugins included in preset
+      //       builtinPluginName: {
+      //         optionName: 'optionValue',
+      //       },
+      //       // or disable plugins
+      //       anotherBuiltinPlugin: false,
+      //     },
+      //   },
+      // },
+      // Enable builtin plugin not included in preset
+      // 'moreBuiltinPlugin',
+      // Enable and configure builtin plugin not included in preset
+      // {
+      //   name: 'manyBuiltInPlugin',
+      //   params: {
+      //     optionName: 'value',
+      //   },
+    },
+  ],
+};