--- a/packages/gestalt/src/icons/text-align-left.svg
+++ b/packages/gestalt/src/icons/text-align-left.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M0 4.5h24v-3H0v3zm0 6h18V7.499H0V10.5zm0 6h24v-3H0v3zm0 6h18v-3H0v3z"/></svg>
+<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24"><path d="M0 4.5h24v-3H0v3zm0 6h18V7.499H0V10.5zm0 6h24v-3H0v3zm0 6h18v-3H0v3z"/></svg>