--- a/packages/gestalt/src/icons/text-align-right.svg
+++ b/packages/gestalt/src/icons/text-align-right.svg
@@ -1,1 +1,1 @@
-<svg height="24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M24 4.5H0v-3h24v3zm0 6H6v-3h18v3zm0 6H0v-3h24v3zm0 6H6v-3h18v3z"/></svg>
+<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24"><path d="M24 4.5H0v-3h24v3zm0 6H6v-3h18v3zm0 6H0v-3h24v3zm0 6H6v-3h18v3z"/></svg>