… packages/gestalt/src/icons/ad-group.svg
… packages/gestalt/src/icons/add-circle.svg
… packages/gestalt/src/icons/add-layout.svg
… packages/gestalt/src/icons/add-pin.svg
… packages/gestalt/src/icons/ads-overview.svg
… packages/gestalt/src/icons/ads-stats.svg
… packages/gestalt/src/icons/alert.svg
… packages/gestalt/src/icons/align-bottom-center.svg
… packages/gestalt/src/icons/align-bottom-left.svg
… packages/gestalt/src/icons/align-bottom-right.svg
… packages/gestalt/src/icons/align-bottom.svg
… packages/gestalt/src/icons/align-middle.svg
… packages/gestalt/src/icons/align-top-center.svg
… packages/gestalt/src/icons/align-top-left.svg
… packages/gestalt/src/icons/align-top-right.svg
… packages/gestalt/src/icons/align-top.svg
… packages/gestalt/src/icons/android-share.svg
… packages/gestalt/src/icons/angled-pin.svg
… packages/gestalt/src/icons/apps.svg
… packages/gestalt/src/icons/arrow-back.svg
… packages/gestalt/src/icons/arrow-circle-down.svg
… packages/gestalt/src/icons/arrow-circle-forward.svg
… packages/gestalt/src/icons/arrow-circle-up.svg
… packages/gestalt/src/icons/arrow-down.svg
… packages/gestalt/src/icons/arrow-end.svg
… packages/gestalt/src/icons/arrow-forward.svg
… packages/gestalt/src/icons/arrow-start.svg
… packages/gestalt/src/icons/arrow-up-right.svg
… packages/gestalt/src/icons/arrow-up.svg
… packages/gestalt/src/icons/bell.svg