### Summary

Install  `npm -g install svgo`

Run `svgo -f packages/gestalt/src/icons --config=packages/gestalt/src/icons/svgo.config.js`


svgo.config.js uses basically the default optimization config. Added `sortAttrs` which sort element attributes for better compression. This is a minor optimization for our icons as we enforce a file structure pattern.

No other plugins seemed to have any positive optimization so they were not included, probably due to the fact we enforce a file structure pattern.

Significant optimizations:

**angled-pin.svg:
Done in 2 ms!
0.588 KiB - 25.4% = 0.438 KiB**

**android-share.svg:
Done in 2 ms!
0.451 KiB - 32.3% = 0.306 KiB**

**arrow-up-right.svg:
Done in 1 ms!
0.446 KiB - 41.8% = 0.26 KiB**

**mute.svg:
Done in 1 ms!
0.639 KiB - 34.9% = 0.416 KiB**

** Optimizations with size increase:**

check-circle.svg:
Done in 1 ms!
0.231 KiB + 3.8% = 0.24 KiB

code.svg:
Done in 1 ms!
0.532 KiB + 5.5% = 0.562 KiB

conversion-tag.svg:
Done in 0 ms!
0.381 KiB + 1.5% = 0.387 KiB

credit-card.svg:
Done in 0 ms!
0.288 KiB + 3.7% = 0.299 KiB

dash.svg:
Done in 1 ms!
0.128 KiB + 3.1% = 0.132 KiB

drag-drop.svg:
Done in 1 ms!
0.363 KiB + 3.2% = 0.375 KiB

report.svg:
Done in 1 ms!
0.245 KiB + 4% = 0.255 KiB

story-pin.svg:
Done in 1 ms!
0.263 KiB + 4.1% = 0.273 KiB

workflow-status-canceled.svg:
Done in 1 ms!
0.368 KiB + 1.6% = 0.374 KiB

sparkle.svg:
Done in 1 ms!
0.462 KiB + 2.5% = 0.474 KiB

**Neglectable optimizations:**

add-circle.svg:
Done in 21 ms!
0.276 KiB - 0.4% = 0.275 KiB

ad.svg:
Done in 1 ms!
0.166 KiB - 0.6% = 0.165 KiB

ad-group.svg:
Done in 3 ms!
0.332 KiB - 0.3% = 0.331 KiB

ads-stats.svg:
Done in 2 ms!
0.451 KiB - 0.2% = 0.45 KiB

add.svg:
Done in 1 ms!
0.158 KiB - 0.6% = 0.157 KiB

ads-overview.svg:
Done in 6 ms!
0.481 KiB - 0.2% = 0.48 KiB

add-pin.svg:
Done in 2 ms!
0.356 KiB - 0.3% = 0.355 KiB

alert.svg:
Done in 2 ms!
0.373 KiB - 0.5% = 0.371 KiB

align-top-center.svg:
Done in 1 ms!
0.334 KiB - 0.3% = 0.333 KiB

align-bottom.svg:
Done in 2 ms!
0.274 KiB - 0.4% = 0.273 KiB

align-top-left.svg:
Done in 1 ms!
0.306 KiB - 0.3% = 0.305 KiB

align-top-right.svg:
Done in 2 ms!
0.304 KiB - 0.3% = 0.303 KiB

align-bottom-left.svg:
Done in 1 ms!
0.308 KiB - 0.3% = 0.307 KiB

align-top.svg:
Done in 1 ms!
0.278 KiB - 0.4% = 0.277 KiB

apps.svg:
Done in 2 ms!
0.366 KiB - 0.3% = 0.365 KiB

arrow-back.svg:
Done in 1 ms!
0.23 KiB - 0.4% = 0.229 KiB

arrow-circle-down.svg:
Done in 5 ms!
0.314 KiB - 0.3% = 0.313 KiB

arrow-circle-forward.svg:
Done in 1 ms!
0.315 KiB - 0.3% = 0.314 KiB

arrow-circle-up.svg:
Done in 1 ms!
0.312 KiB - 0.3% = 0.311 KiB

arrow-down.svg:
Done in 1 ms!
0.214 KiB - 0.9% = 0.212 KiB

arrow-end.svg:
Done in 1 ms!
0.232 KiB - 0.4% = 0.231 KiB

arrow-start.svg:
Done in 0 ms!
0.229 KiB - 0.4% = 0.229 KiB

arrow-up.svg:
Done in 1 ms!
0.24 KiB - 0.4% = 0.239 KiB

calendar.svg:
Done in 0 ms!
0.25 KiB - 0.4% = 0.249 KiB

camera-roll.svg:
Done in 0 ms!
0.349 KiB - 0.3% = 0.348 KiB

bell.svg:
Done in 0 ms!
0.223 KiB - 0.4% = 0.222 KiB

add-layout.svg:
Done in 2 ms!
0.505 KiB - 0.2% = 0.504 KiB

arrow-forward.svg:
Done in 0 ms!
0.234 KiB - 0.4% = 0.233 KiB

camera.svg:
Done in 1 ms!
0.405 KiB - 0.2% = 0.404 KiB

circle-outline.svg:
Done in 0 ms!
0.245 KiB - 0.4% = 0.244 KiB

clock.svg:
Done in 0 ms!
0.291 KiB - 0.3% = 0.29 KiB

cog.svg:
Done in 0 ms!
0.164 KiB - 0.6% = 0.163 KiB

compass.svg:
Done in 1 ms!
0.404 KiB - 0.2% = 0.403 KiB

color-picker.svg:
Done in 1 ms!
0.352 KiB - 0.3% = 0.351 KiB

compose.svg:
Done in 1 ms!
0.299 KiB - 0.3% = 0.298 KiB

crop.svg:
Done in 0 ms!
0.295 KiB - 0.3% = 0.294 KiB

directional-arrow-left.svg:
Done in 1 ms!
0.195 KiB - 0.5% = 0.194 KiB

directional-arrow-right.svg:
Done in 0 ms!
0.2 KiB - 0.5% = 0.199 KiB

download.svg:
Done in 1 ms!
0.295 KiB - 0.3% = 0.294 KiB

duplicate.svg:
Done in 1 ms!
0.504 KiB - 0.2% = 0.503 KiB

edit.svg:
Done in 1 ms!
0.25 KiB - 0.8% = 0.248 KiB

ellipsis-circle-outline.svg:
Done in 1 ms!
0.452 KiB - 0.2% = 0.451 KiB

ellipsis.svg:
Done in 2 ms!
0.248 KiB - 0.4% = 0.247 KiB

envelope.svg:
Done in 1 ms!
0.315 KiB - 0.6% = 0.313 KiB

eye-hide.svg:
Done in 1 ms!
0.561 KiB - 0.3% = 0.559 KiB


eye.svg:
Done in 1 ms!
0.427 KiB - 36.2% = 0.272 KiB

face-sad.svg:
Done in 1 ms!
0.638 KiB - 0.2% = 0.637 KiB

face-happy.svg:
Done in 1 ms!
0.467 KiB - 1.5% = 0.46 KiB

face-smiley.svg:
Done in 1 ms!
0.413 KiB - 0.2% = 0.412 KiB

file-unknown.svg:
Done in 1 ms!
0.666 KiB - 0.1% = 0.665 KiB

facebook.svg:
Done in 1 ms!
0.259 KiB - 0.8% = 0.257 KiB

fill-opaque.svg:
Done in 1 ms!
0.336 KiB - 0.3% = 0.335 KiB

canonical-pin.svg:
Done in 1 ms!
0.416 KiB - 0.2% = 0.415 KiB

fill-transparent.svg:
Done in 1 ms!
0.435 KiB - 0.2% = 0.434 KiB

flag.svg:
Done in 0 ms!
0.133 KiB - 0.7% = 0.132 KiB

filter.svg:
Done in 1 ms!
0.541 KiB - 0.2% = 0.54 KiB

check.svg:
Done in 0 ms!
0.231 KiB - 0.8% = 0.229 KiB

flash.svg:
Done in 1 ms!
0.124 KiB - 0.8% = 0.123 KiB

cancel.svg:
Done in 1 ms!
0.384 KiB - 0.3% = 0.383 KiB

folder.svg:
Done in 1 ms!
0.207 KiB - 0.5% = 0.206 KiB

flip-horizontal.svg:
Done in 1 ms!
0.265 KiB - 0.4% = 0.264 KiB

flashlight.svg:
Done in 3 ms!
0.583 KiB - 0.2% = 0.582 KiB

gmail.svg:
Done in 0 ms!
0.151 KiB - 0.6% = 0.15 KiB

globe-checked.svg:
Done in 2 ms!
0.594 KiB - 0.2% = 0.593 KiB

google-plus.svg:
Done in 1 ms!
0.351 KiB - 0.3% = 0.35 KiB

graph-bar.svg:
Done in 1 ms!
0.133 KiB - 0.7% = 0.132 KiB

hand-pointing.svg:
Done in 2 ms!
1.237 KiB - 30% = 0.866 KiB

handle.svg:
Done in 0 ms!
0.192 KiB - 0.5% = 0.191 KiB

heart-outline.svg:
Done in 2 ms!
0.603 KiB - 0.2% = 0.602 KiB

heart-broken.svg:
Done in 1 ms!
0.487 KiB - 0.2% = 0.486 KiB

impressum.svg:
Done in 1 ms!
0.42 KiB - 0.2% = 0.419 KiB

heart.svg:
Done in 1 ms!
0.278 KiB - 0.7% = 0.276 KiB

info-circle.svg:
Done in 1 ms!
0.247 KiB - 0.4% = 0.246 KiB

knoop.svg:
Done in 3 ms!
0.437 KiB - 0.2% = 0.436 KiB

key.svg:
Done in 1 ms!
0.225 KiB - 0.4% = 0.224 KiB

lightbulb.svg:
Done in 9 ms!
0.29 KiB - 0.3% = 0.289 KiB

clear.svg:
Done in 1 ms!
0.403 KiB - 0.2% = 0.402 KiB

lightning-bolt-circle.svg:
Done in 1 ms!
0.291 KiB - 0.7% = 0.289 KiB

history.svg:
Done in 1 ms!
0.689 KiB - 30.2% = 0.481 KiB

link.svg:
Done in 1 ms!
0.509 KiB - 0.4% = 0.507 KiB

lock.svg:
Done in 1 ms!
0.247 KiB - 0.4% = 0.246 KiB

location.svg:
Done in 1 ms!
0.259 KiB - 0.4% = 0.258 KiB

logo-large.svg:
Done in 1 ms!
0.464 KiB - 0.4% = 0.462 KiB

logo-small.svg:
Done in 1 ms!
0.438 KiB - 0.2% = 0.438 KiB

margins-large.svg:
Done in 0 ms!
0.354 KiB - 0.6% = 0.353 KiB

logout.svg:
Done in 1 ms!
0.5 KiB - 0.2% = 0.499 KiB

gif.svg:
Done in 3 ms!
0.909 KiB - 0.1% = 0.908 KiB

megaphone.svg:
Done in 0 ms!
0.329 KiB - 0.3% = 0.328 KiB

margins-small.svg:
Done in 1 ms!
0.441 KiB - 0.4% = 0.439 KiB

minimize.svg:
Done in 1 ms!
0.375 KiB - 0.3% = 0.374 KiB

menu.svg:
Done in 1 ms!
0.393 KiB - 0.2% = 0.392 KiB

flip-vertical.svg:
Done in 1 ms!
0.275 KiB + 0.4% = 0.276 KiB

margins-medium.svg:
Done in 0 ms!
0.438 KiB - 0.4% = 0.437 KiB

move.svg:
Done in 2 ms!
0.746 KiB - 0.1% = 0.745 KiB

pause.svg:
Done in 1 ms!
0.229 KiB - 0.4% = 0.229 KiB

overlay-text.svg:
Done in 0 ms!
0.388 KiB - 0.3% = 0.387 KiB

people.svg:
Done in 0 ms!
0.286 KiB - 0.3% = 0.285 KiB

person-add.svg:
Done in 1 ms!
0.301 KiB - 0.3% = 0.3 KiB

phone.svg:
Done in 1 ms!
0.246 KiB + 4.8% = 0.258 KiB

person.svg:
Done in 0 ms!
0.181 KiB - 0.5% = 0.18 KiB

pin-hide.svg:
Done in 1 ms!
0.498 KiB - 0.2% = 0.497 KiB

pin.svg:
Done in 1 ms!
0.214 KiB - 0.5% = 0.213 KiB

pinterest.svg:
Done in 1 ms!
0.801 KiB - 0.1% = 0.8 KiB

play.svg:
Done in 2 ms!
0.179 KiB - 0.5% = 0.178 KiB

question-mark.svg:
Done in 2 ms!
0.443 KiB - 0.2% = 0.442 KiB

refresh.svg:
Done in 0 ms!
0.346 KiB - 0.3% = 0.345 KiB

remove.svg:
Done in 1 ms!
0.208 KiB - 0.5% = 0.207 KiB

shopping-bag.svg:
Done in 1 ms!
0.251 KiB + 5.8% = 0.266 KiB


replace.svg:
Done in 1 ms!
0.639 KiB - 0.2% = 0.638 KiB

reorder-images.svg:
Done in 6 ms!
0.545 KiB - 0.2% = 0.544 KiB

rotate.svg:
Done in 1 ms!
0.507 KiB - 0% = 0.507 KiB

scale.svg:
Done in 1 ms!
0.42 KiB - 0.2% = 0.419 KiB

send.svg:
Done in 0 ms!
0.357 KiB - 0.5% = 0.355 KiB

security.svg:
Done in 0 ms!
0.185 KiB - 0.5% = 0.184 KiB

share.svg:
Done in 1 ms!
0.346 KiB - 0.3% = 0.345 KiB

smiley-outline.svg:
Done in 2 ms!
0.651 KiB - 0.1% = 0.65 KiB

search.svg:
Done in 1 ms!
0.296 KiB - 0.7% = 0.294 KiB

smiley.svg:
Done in 1 ms!
0.486 KiB - 0.2% = 0.485 KiB

sort-ascending.svg:
Done in 0 ms!
0.195 KiB - 0.5% = 0.194 KiB

maximize.svg:
Done in 1 ms!
0.375 KiB - 0.3% = 0.374 KiB

sound.svg:
Done in 6 ms!
0.589 KiB - 0.2% = 0.588 KiB

speech-ellipsis.svg:
Done in 1 ms!
0.364 KiB - 0.3% = 0.363 KiB

speech.svg:
Done in 1 ms!
0.23 KiB - 0.4% = 0.229 KiB

star-half.svg:
Done in 0 ms!
0.272 KiB - 0.4% = 0.271 KiB

star.svg:
Done in 1 ms!
0.413 KiB - 0.2% = 0.412 KiB

switch-account.svg:
Done in 1 ms!
0.451 KiB - 0.2% = 0.45 KiB

tag.svg:
Done in 1 ms!
0.284 KiB - 0.3% = 0.283 KiB

terms.svg:
Done in 1 ms!
0.289 KiB - 0.3% = 0.288 KiB

text-align-center.svg:
Done in 0 ms!
0.147 KiB - 0.7% = 0.146 KiB

text-align-left.svg:
Done in 1 ms!
0.146 KiB - 0.7% = 0.146 KiB

text-align-right.svg:
Done in 0 ms!
0.142 KiB - 0.7% = 0.141 KiB

text-all-caps.svg:
Done in 1 ms!
0.654 KiB - 0.1% = 0.653 KiB

text-extra-small.svg:
Done in 1 ms!
0.435 KiB - 0.2% = 0.434 KiB

text-large.svg:
Done in 1 ms!
0.462 KiB - 0.2% = 0.461 KiB

text-medium.svg:
Done in 1 ms!
0.436 KiB - 0.2% = 0.435 KiB

text-sentence-case.svg:
Done in 1 ms!
0.688 KiB - 0.1% = 0.688 KiB

globe.svg:
Done in 1 ms!
0.609 KiB - 0.2% = 0.608 KiB

text-small.svg:
Done in 0 ms!
0.438 KiB - 0.2% = 0.438 KiB

text-size.svg:
Done in 1 ms!
0.563 KiB - 0.3% = 0.562 KiB

text-spacing.svg:
Done in 1 ms!
0.701 KiB - 0.1% = 0.7 KiB

twitter.svg:
Done in 1 ms!
0.566 KiB - 0.5% = 0.563 KiB

sort-descending.svg:
Done in 0 ms!
0.199 KiB - 0.5% = 0.198 KiB

video-camera.svg:
Done in 1 ms!
0.249 KiB - 0.8% = 0.247 KiB

view-type-default.svg:
Done in 0 ms!
0.169 KiB - 0.6% = 0.168 KiB

view-type-dense.svg:
Done in 5 ms!
0.257 KiB - 0.4% = 0.256 KiB

view-type-list.svg:
Done in 1 ms!
0.135 KiB - 0.7% = 0.134 KiB

visit.svg:
Done in 1 ms!
0.367 KiB - 0.3% = 0.366 KiB

workflow-status-all.svg:
Done in 1 ms!
0.305 KiB - 0.3% = 0.304 KiB

workflow-status-halted.svg:
Done in 0 ms!
0.296 KiB - 0.3% = 0.295 KiB

workflow-status-in-progress.svg:
Done in 1 ms!
0.164 KiB - 0.6% = 0.163 KiB

text-line-height.svg:
Done in 0 ms!
0.266 KiB - 0.7% = 0.264 KiB

workflow-status-problem.svg:
Done in 2 ms!
0.429 KiB - 0.2% = 0.428 KiB

workflow-status-unstarted.svg:
Done in 1 ms!
0.263 KiB - 0.4% = 0.262 KiB

trash-can.svg:
Done in 1 ms!
0.266 KiB - 0.4% = 0.265 KiB

align-bottom-center.svg:
Done in 2 ms!
0.34 KiB - 0.3% = 0.339 KiB

workflow-status-warning.svg:
Done in 4 ms!
0.319 KiB - 0.6% = 0.317 KiB

align-bottom-right.svg:
Done in 2 ms!
0.306 KiB - 0.3% = 0.305 KiB

workflow-status-ok.svg:
Done in 0 ms!
0.263 KiB - 0.7% = 0.261 KiB

align-middle.svg:
Done in 2 ms!
0.344 KiB - 0.3% = 0.343 KiB

<!--
What is the purpose of this PR?

Have you [formatted the PR title](https://github.com/pinterest/gestalt/#releasing)? `ComponentName: Description`
-->

### Links

- [Jira](link to Jira ticket(s))
- [TDD](link to Paper doc)
- [Figma](link to Figma file)

### Checklist

- [ ] Added unit and Flow Tests
- [ ] Added documentation + accessibility tests
- [ ] Verified accessibility: keyboard & screen reader interaction
- [ ] Checked dark mode, responsiveness, and right-to-left support
- [ ] Checked stakeholder feedback (e.g. Gestalt designers)